fix 64 bit compilation warning
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/branches/WX_2_6_BRANCH@44955 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/appbase.cpp
@@ -721,7 +721,8 @@ static wxString GetAssertStackTrace()
     protected:
         virtual void OnStackFrame(const wxStackFrame& frame)
         {
-            m_stackTrace << wxString::Format(_T("[%02d] "), frame.GetLevel());
+            m_stackTrace << wxString::Format(_T("[%02lu] "),
+                                             (unsigned long)frame.GetLevel());
 
             wxString name = frame.GetName();
             if ( !name.empty() )
